Update Undo handle

+++ stategameready.cpp
@@ -96,7 +96,7 @@ bool StateGameReady::GameRequestHandle(const string &msg)
             //Notify the others that should update user's info
             requestReply.set_status(2);
             requestReply.SerializeToString(&data);
-            stateMachine->currChessBoard->BroadCastMsg(MSG_REQUEST_PLAY_REPLY, data, stateMachine->locate);
+            stateMachine->currChessBoard->BroadCastMsg(MSG_REQUEST_PLAY_REPLY, data, (int)stateMachine->locate);
         }
     }
